MDL-27156 repositories: altered filepicker.php so that users with the ability to ignore course file size restrictions are not incorrectly restricted

+++ lib/form/filepicker.php
@@ -207,7 +207,10 @@ class MoodleQuickForm_filepicker extends HTML_QuickForm_input {
             $usercontext = get_context_instance(CONTEXT_USER, $USER->id);
             if ($files = $fs->get_area_files($usercontext->id, 'user', 'draft', $draftitemid, 'id DESC', false)) {
                 $file = array_shift($files);
-                if ($this->_options['maxbytes'] and $file->get_filesize() > $this->_options['maxbytes']) {
+                if ($this->_options['maxbytes']
+                    and $this->_options['maxbytes'] !== USER_CAN_IGNORE_FILE_SIZE_LIMITS
+                    and $file->get_filesize() > $this->_options['maxbytes']) {
+
                     // bad luck, somebody tries to sneak in oversized file
                     $file->delete();
                 }
